fix(hs): update chakra Input styles
