+++ packages/headless-styles/src/chakra.ts
@@ -4,7 +4,7 @@ import { ChakraErrorMessage } from './components/ErrorMessage/errorMessageJS'
 import { ChakraFormControl } from './components/FormControl/formControlJS'
 import { ChakraFormLabel } from './components/FormLabel/formLabelJS'
 import { ChakraIcon } from './components/Icon/iconJS'
-import { ChakraInput } from './components/Input/inputJS'
+import { ChakraInput } from './components/Input/chakraInput'
 import { ChakraProgress } from './components/Progress/progressJS'
 import { ChakraRadio } from './components/Radio/radioJS'
 import { ChakraSkeleton } from './components/Skeleton/skeletonJS'
