+++ packages/headless-styles/src/components/Input/inputJS.ts
@@ -3,32 +3,11 @@ import { createInputOptions, getDefaultInputOptions } from './shared'
 import styles from './generated/InputCSS.module'
 import type { InputOptions } from './types'
 
-const baseInputStyles = styles.inputBase
-
-export const ChakraInput = {
-  baseStyle: baseInputStyles,
-  sizes: {
-    m: {
-      ...baseInputStyles,
-      ...styles.mInputSize,
-    },
-    l: {
-      ...baseInputStyles,
-    },
-  },
-  variants: {
-    outline: baseInputStyles,
-    filled: baseInputStyles,
-    flushed: baseInputStyles,
-    unstyled: baseInputStyles,
-  },
-}
-
 export function getJSInputProps(options?: InputOptions) {
   const defaultOptions = getDefaultInputOptions(options)
   const inputProps = createInputOptions(defaultOptions)
   const jsStyles = {
-    ...baseInputStyles,
+    ...styles.inputBase,
     ...styles[`${defaultOptions.size}InputSize`],
     ['&::placeholder']: {
       ...styles['']['&::placeholder'],
